+++ cmd/root.go
@@ -3,20 +3,28 @@ package cmd
 import (
 	"gover/internal/utils"
 	"gover/pkg/config"
+	"log"
 
 	"github.com/spf13/cobra"
 )
 
-var variables *config.Variables
+var cfg *config.Config
 
-func NewRootCmd(v *config.Variables) *cobra.Command {
+func NewRootCmd(c *config.Config) *cobra.Command {
 	cmd := &cobra.Command{
 		Use:   "gover",
 		Short: "gover is a tool to get project version",
 		Long:  `gover gets the project vertsion using a combination of a VERSION file as well as GitLab CI/CD variables`,
 		PersistentPreRun: func(cmd *cobra.Command, args []string) {
 			// load variables
-			variables = v
+			cfg = c
+			if cfg.Debug {
+				log.Println("Checking variables")
+			}
+			ok, mv := cfg.CheckVariables()
+			if !ok {
+				log.Fatalf("Missing variables: %v", mv)
+			}
 		},
 		Run: func(cmd *cobra.Command, args []string) {
 			cmd.Help()
@@ -32,7 +40,7 @@ func newVersionCmd() *cobra.Command {
 		Short: "Print the version number of gover",
 		Long:  `All software has versions. This is gover's`,
 		Run: func(cmd *cobra.Command, args []string) {
-			cmd.Println(utils.GetVersion(variables))
+			cmd.Println(utils.GetVersion(cfg))
 		},
 	}
 	return cmd